move log and tasks namespaces to ctBrowse_lib
