+++ app/LabelImageCache.h
@@ -8,8 +8,8 @@
 #pragma once
 
 #include "App.h"
-#include "tasks.h"
 
+#include <ctb/tasks/tasks.h>
 #include <wx/image.h>
 
 #include <expected>
