+++ include/ctb/ctb.h
@@ -11,6 +11,7 @@
 #include "ctb/concepts.h"
 #include "ctb/common_types.h"
 #include "ctb/Error.h"
+#include "ctb/log.h"
 
 namespace ctb
 {
