+++ lib/tasks.cpp
@@ -1,7 +1,7 @@
-#include "App.h"
-#include "tasks.h"
+#include "ctb/tasks/tasks.h"
+#include "ctb/utility.h"
+#include "ctb/utility_http.h"
 
-#include <ctb/utility_http.h>
 #include <cpr/cpr.h>
