+++ vcpkg.json
@@ -22,5 +22,5 @@
     },
     "spdlog"
   ],
-  "builtin-baseline": "b328539dd9c022e3f11b267ef68f11ad3a85eaf8"
+  "builtin-baseline": "96d5fb3de135b86d7222c53f2352ca92827a156b"
 }
